Check for deeper getString() call paths
Allow calling of this.gettextCatalog.getString(), not just direct access of gettextCatalog.getString().

We were injecting gettextCatalog in our controller and storing it in the class, without this change would have to reassign it to a var each time we want to access getString(). This change detects longer call paths, as long as the last part is gettextCatalog.getString().