Convert horizontal ellipsis in msgids from html-entities to characters
angular-gettext-extract extracts &hellip; as a html-entity into .po-files, but browsers automatically convert &hellip; to …

This means that your translations-file may have a msgid "dots&hellip;", but the translate-directive will read "dots…" from a html-element and try to find that as a msgid (and find nothing).

To get around this we convert any &hellip; to … when compiling from a .po-file.