Add plural marker name support to extraction.
Add support for specifying the options `markerNamePlural` and/or `markerNamesPlural`, which will be used to extract strings from matching function calls. This allows using a regular function (as opposed to a method) for pluralizable catalog messages.

If the options are not specified, the extractor continues to work as before. Includes unit tests to cover the new behavior.